refactor(step1):  Separator
Separators 안의 String을 Separator로 분리
