+++ src/main/java/stringpluscalculator/Separators.java
@@ -1,22 +1,26 @@
 package stringpluscalculator;
 
 import java.util.Arrays;
+import java.util.List;
+import java.util.stream.Collectors;
 
 public class Separators {
     private static final String CUSTOM_SEPARATORS_PREFIX = "//";
     private static final String CUSTOM_SEPARATORS_SUFFIX = "\\n";
     private static final String SPLIT_DELIMITER = "|";
 
-    private String[] separators = {":", ","};
+    private List<Separator> separators = Arrays.asList(new Separator(":"), new Separator(","));
 
     public Separators(String input) {
         if (hasCustomSeparator(input)) {
-            this.separators = new String[]{searchCustomSeparator(input)};
+            this.separators = Arrays.asList(new Separator(searchCustomSeparator(input)));
         }
     }
 
     public String getSplitSeparator() {
-        return String.join(SPLIT_DELIMITER, this.separators);
+        return separators.stream()
+                .map(Separator::getSeparator)
+                .collect(Collectors.joining(SPLIT_DELIMITER));
     }
 
     public static String getInputWithoutCustomSeparator(String input) {
@@ -38,18 +42,5 @@ public class Separators {
         int suffixIndex = input.indexOf(CUSTOM_SEPARATORS_SUFFIX);
 
         return input.substring(CUSTOM_SEPARATORS_PREFIX.length(), suffixIndex);
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        Separators that = (Separators) o;
-        return Arrays.equals(separators, that.separators);
-    }
-
-    @Override
-    public int hashCode() {
-        return Arrays.hashCode(separators);
     }
 }
